Pdf scrapper works!
Successful intercepted fisher api and got the pdf, still needs to be integrated and parsed for correct information. Need to add a cookie generator as well as a random 3-7sec timeout between requests to not get IP banned lol. Also need to actually extract the data I need from the pdf and change code to have the cas number and search be dynamic, not static with acetone.

CSV export still needs be corrected, writes file properly but does not export data correctly - fix object formating here.

Still need to figure out how to import the data we have and store it.

Probably about 70% done with project :).